Move const from csdev to bsmp.constants
