+++ siriuspy/siriuspy/pwrsupply/pructrl/udc.py
@@ -4,7 +4,8 @@ import time as _time
 import numpy as _np
 
 from . import prucparms as _prucparms
-from ..csdev import UDC_MAX_NR_DEV as _UDC_MAX_NR_DEV
+
+from ..bsmp.constants import UDC_MAX_NR_DEV as _UDC_MAX_NR_DEV
 from ..bsmp.factory import PSBSMPFactory as _PSBSMPFactory
