+++ siriuspy/tests/pwrsupply/test_csdev.py
@@ -19,7 +19,6 @@ PUB_INTERFACE = (
     'DEFAULT_WFM_OTHERS',
     'DEFAULT_WFM',
     'PSSOFB_MAX_NR_UDC',
-    'UDC_MAX_NR_DEV',
     'DEFAULT_SIGGEN_CONFIG',
     'PS_CURRENT_PRECISION',
     'PU_VOLTAGE_PRECISION',
